requirements-voice.txt
pandas==2.0.3
numpy==1.24.3
httpx==0.24.1
openai==1.3.0
python-dotenv==1.0.0

# STT - Speech to Text
openai-whisper==20231106
faster-whisper==0.10.0

# TTS - Text to Speech (choose at least one)
TTS==0.22.0
# git+https://github.com/hexgrad/Kokoro.git  # Uncomment for Kokoro

# Clinical NLP
transformers==4.34.0
torch==2.0.1
spacy==3.7.2
# python -m spacy download en_core_sci_md  # Run after install

# Audio Processing
librosa==0.10.0
soundfile==0.12.1
pydub==0.25.1
scipy==1.11.3

# EHR and FHIR
fhirclient==4.2.1
aiohttp==3.9.0

# Async and concurrent
asyncio-contextmanager==1.0.0
aiofiles==23.2.1

# Utilities
requests==2.31.0
pyyaml==6.0.1
pydantic==2.3.0

# CLI and web (optional)
click==8.1.7
typer==0.9.0
uvicorn==0.23.2
fastapi==0.103.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.9.1
flake8==6.1.0
mypy==1.5.1
